workbench/lcar_lab/training/requirements.txt
google-genai>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
aiohttp>=3.8.0
asyncio>=3.4.3
sqlalchemy>=2.0.0 
r2r==0.1.0
aiolimiter>=0.1.0
scikit-learn>=1.0.0  # For train_test_split
pathlib>=1.0.1  # For Path operations
typing-extensions>=4.0.0  # For typing support